Updated CellularAutomata
Added a extra Graph which all the results will be calculated to. The main graph is updated with the values of this new graph, after all calculations are done.  This avoids the errors you would get if you were to update the values of the main graph as you went, as the new values will influence the values of its neighbours which you don't want happening.