Done the calculate method of CellulaAutomata
The calculate method is finished, but needs testing. It should update the equivalent cell in graph2 from the value of the surrounding equivalent cell in graph.  I now need to implement the step method which will iteraye through all the cells in graph, updating graph2, then copy over the values of graph2 to graph.