Added a way to find out what classes are contained in the grid
Wrote a method Graph::generate_classes which creates and returns a set<int> with values inserted from each position in the grid.  Since it's a set, only those unique values will be present representing the range of classes contained in the grid.  This way the grid can have as may classes as the implementer wants.