Changes are made to make the memory footprint smaller
Have written a programme to find out what the maximum ammount of memory is that can be allocated using malloc.

Am trying to reduce the memory footprint by specifying the number of bytes used by a variable in a c struct via the struct { int c :10; } method. This tells it to use only 10 bits for c.