Modified programme, and transform script. Wrote script to run all this.
Reorganised the programme to accept 3 input files. the first being the
grid definitions, then training data, and finally the test data.

Modified transform_abalone.py to accept a single integer input, which
determines the size of testing data by taking 1/input as the number of
pieces to use, I changed this for the last test run I did so that the
number is the size of the training data instead. 1/input of the data
is training data. Renamed to transform.py.

Wrote a script (run.sh to run first the transform.py with different
input values, then the programme, using these resulting files as its
training and testing data.